split up legacyUtil.js
