+++ creator-node/package.json
@@ -87,7 +87,9 @@
     "@types/bull": "^3.15.8",
     "@types/bunyan": "^1.8.8",
     "@types/chai": "^4.3.3",
+    "@types/eth-sig-util": "^2.1.1",
     "@types/express": "4.17.12",
+    "@types/fs-extra": "^9.0.13",
     "@types/lodash": "^4.14.182",
     "@types/mocha": "^9.1.1",
     "@types/node": "^18.7.9",
