+++ creator-node/src/utils/index.ts
@@ -1,25 +1,27 @@
 import type { ReplicaSet } from './strToReplicaSet'
 import { strToReplicaSet } from './strToReplicaSet'
-import {
-  timeout,
-  getRandomInt,
-  verifySignature,
-  currentNodeShouldHandleTranscode,
-  validateStateForImageDirCIDAndReturnFileUUID,
-  findCIDInNetwork,
-  getAllRegisteredCNodes,
-  getIfAttemptedStateFix,
-  createDirForFile,
-  writeStreamToFileSystem,
-  _streamFileToDiskHelper,
-  runShellCommand,
-  verifyCIDMatchesExpected,
-  EMPTY_FILE_CID
-} from './legacyUtils'
+import { timeout, getRandomInt, verifySignature } from './utils'
 import {
   validateMetadata,
   validateAssociatedWallets
 } from './validateAudiusUserMetadata'
+import {
+  findCIDInNetwork,
+  verifyCIDMatchesExpected,
+  EMPTY_FILE_CID
+} from './cidUtils'
+import {
+  createDirForFile,
+  writeStreamToFileSystem,
+  getIfAttemptedStateFix,
+  validateStateForImageDirCIDAndReturnFileUUID,
+  _streamFileToDiskHelper
+} from './fsUtils'
+import { runShellCommand } from './runShellCommand'
+import {
+  currentNodeShouldHandleTranscode,
+  getAllRegisteredCNodes
+} from './cnodeUtils'
 
 export type { ReplicaSet }
 export {
